refactor(moon): improve error handling and validation
- Replace direct nil checks with validate.IsNil() function
- Use more specific error types instead of generic internal server error
- Add validation logic for SaveTeamMetricDatasource request
- Update DatasourceMetric.ScrapeInterval to int64 type

+++ cmd/palace/internal/data/impl/team_datasource_metric.go
@@ -62,7 +62,7 @@ func (t *teamMetricDatasourceImpl) Update(ctx context.Context, req *bo.SaveTeamM
 		mutation.Remark.Value(req.Remark),
 		mutation.Driver.Value(req.Driver.GetValue()),
 		mutation.Endpoint.Value(req.Endpoint),
-		mutation.ScrapeInterval.Value(int64(req.ScrapeInterval)),
+		mutation.ScrapeInterval.Value(req.ScrapeInterval),
 		mutation.Headers.Value(crypto.NewObject(req.Headers)),
 		mutation.QueryMethod.Value(req.QueryMethod.GetValue()),
 		mutation.Extra.Value(crypto.NewObject(req.Extra)),
